displays graph data from database
